dashboard/requirements.txt
pyspark==3.4.1
delta-spark==2.4.0
streamlit==1.35.0
pandas
